Proof of concept for a complex intermingling of Zig and C code in a hybrid project
Summary:
Ideally trying to recreate this graph:
```
flowchart
    shared{"Shared"}
    root["Root"]
    baz("Baz")
    c("C")

    shared--referenced by-->baz
    c & shared--referenced by-->root
    baz--referenced by-->c


    c -.call baz().-> baz
    root -.call init() / call get().-> shared
    root -.call foo().-> c
    baz -.call get().-> shared
```

The complexity here comes with trying to work in a globally shared singleton from zig that is both exposed to other zig modules as well as c code directly compiled into the module. There's a few things here to battle with, and it doesn't really come out as very ziggish but that is probably to be expected.

Test Plan: